fix #22 Adjusted point spacing and Descriptive Test methods
LIC10 was originally misinterpreted. Point offsets have been adjusted to align with the description of the assignment.

Test methods have been changed to adhere to the changes in the method.

Some error handling has been introduced into #12 LIC10. The method will return false but does not throw any errors if the input is not in line with description limitations.

fixed minor spelling errors.